% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmex}
\alias{denmex}
\title{denmex}
\format{
A data.frame object with 5156780 rows and 3 variables:
\describe{
  \item{CVEGEO}{is id of municipality}
  \item{date}{is the date}
  \item{n}{dengue cases}
  ...
}
}
\source{
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}
}
\usage{
denmex
}
\description{
denmex: daily dengue dataset for each municipality in Mexico
}
\details{
climex provides the daily dengue data for each municipality in Mexico, downloaded from the [datos abiertod DGE](https://www.gob.mx/salud/documentos/datos-abiertos-152127). The time series covers the period from January 1, 2020, to Octuber 22, 2025 (two weeks delate).
}
\keyword{datasets}
