% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popmex}
\alias{popmex}
\title{popmex}
\format{
A data.frame object with 2469 rows and 3 variables:
\describe{
  \item{CVEGEO}{id of municipality}
  \item{nom_mun}{name of municipality}
  \item{population}{population}
  ...
}
}
\usage{
popmex
}
\description{
popmex: population dataset for each municipality in Mexico
}
\details{
popmex provides population data for each municipality in Mexico, downloaded from the [INEGI]().
}
\keyword{datasets}
